+++ app/src/main/java/com/mishaismenska/hackatonrsschoolapp/domain/interfaces/SetUserWeightUseCase.kt
@@ -1,5 +1,5 @@
 package com.mishaismenska.hackatonrsschoolapp.domain.interfaces
 
 interface SetUserWeightUseCase {
-    suspend fun setUserWeight(weight: Int)
+    suspend fun setUserWeight(weight: Double)
 }
